Revert "Make holiday stops work with new T3 product"
