--- handlers/holiday-stop-api/src/test/resources/TierThreeAccount.json
@@ -1,123 +1,0 @@
-{
-  "basicInfo": {
-    "id": "8ad08f069010dd31019011e437574822",
-    "name": "001UD000006jIF3YAM",
-    "accountNumber": "A00911450",
-    "notes": null,
-    "status": "Active",
-    "crmId": "001UD000006jIF3YAM",
-    "batch": "Batch1",
-    "invoiceTemplateId": "2c92c0f849369b8801493bf7db7e450e",
-    "communicationProfileId": null,
-    "profileNumber": null,
-    "purchaseOrderNumber": null,
-    "customerServiceRepName": null,
-    "ScrubbedOn__c": null,
-    "IdentityId__c": "200264404",
-    "sfContactId__c": "003UD00000BDAMbYAP",
-    "CCURN__c": null,
-    "NonStandardDataReason__c": null,
-    "Scrubbed__c": false,
-    "ProcessingAdvice__c": "NoAdvice",
-    "CreatedRequestId__c": "063a4ed5-c059-df66-0000-00000000065d",
-    "RetryStatus__c": null,
-    "salesRep": null,
-    "parentId": null,
-    "creditMemoTemplateId": null,
-    "debitMemoTemplateId": null,
-    "summaryStatementTemplateId": null,
-    "sequenceSetId": null
-  },
-  "billingAndPayment": {
-    "billCycleDay": 28,
-    "currency": "GBP",
-    "paymentTerm": "Due Upon Receipt",
-    "paymentGateway": "Stripe PaymentIntents GNM Membership",
-    "paymentGatewayNumber": "PG-00000013",
-    "defaultPaymentMethodId": "8ad08f069010dd310190269688ea41dd",
-    "invoiceDeliveryPrefsPrint": false,
-    "invoiceDeliveryPrefsEmail": false,
-    "autoPay": true,
-    "additionalEmailAddresses": [],
-    "paymentMethodCascadingConsent": false
-  },
-  "metrics": {
-    "balance": 0.000000000,
-    "currency": "GBP",
-    "totalInvoiceBalance": 0.000000000,
-    "creditBalance": 0.000000000,
-    "totalDebitMemoBalance": 0.000000000,
-    "unappliedPaymentAmount": 0.000000000,
-    "unappliedCreditMemoAmount": 0.000000000,
-    "contractedMrr": 25.000000000
-  },
-  "billToContact": {
-    "accountId": "8ad08f069010dd31019011e437574822",
-    "accountNumber": "A00911450",
-    "address1": "KP",
-    "address2": "",
-    "city": "London",
-    "country": "United Kingdom",
-    "county": null,
-    "fax": null,
-    "firstName": "R",
-    "homePhone": null,
-    "id": "8ad08f069010dd31019011e437784825",
-    "lastName": "B",
-    "mobilePhone": null,
-    "nickname": null,
-    "otherPhone": null,
-    "otherPhoneType": null,
-    "personalEmail": null,
-    "state": null,
-    "taxRegion": null,
-    "workEmail": "rupert.bates+t3@observer.co.uk",
-    "workPhone": null,
-    "zipCode": "N19GU",
-    "Company_Name__c": null,
-    "SpecialDeliveryInstructions__c": null,
-    "Title__c": null,
-    "contactDescription": null
-  },
-  "soldToContact": {
-    "accountId": "8ad08f069010dd31019011e437574822",
-    "accountNumber": "A00911450",
-    "address1": "KP",
-    "address2": "",
-    "city": "London",
-    "country": "United Kingdom",
-    "county": null,
-    "fax": null,
-    "firstName": "R",
-    "homePhone": null,
-    "id": "8ad08f069010dd31019011e437854827",
-    "lastName": "B",
-    "mobilePhone": null,
-    "nickname": null,
-    "otherPhone": null,
-    "otherPhoneType": null,
-    "personalEmail": null,
-    "state": null,
-    "taxRegion": null,
-    "workEmail": "rupert.bates+t3@observer.co.uk",
-    "workPhone": null,
-    "zipCode": "N19GU",
-    "Company_Name__c": null,
-    "SpecialDeliveryInstructions__c": null,
-    "Title__c": null,
-    "contactDescription": null
-  },
-  "taxInfo": {
-    "exemptStatus": "No",
-    "exemptCertificateId": null,
-    "exemptCertificateType": null,
-    "exemptIssuingJurisdiction": null,
-    "exemptEffectiveDate": null,
-    "exemptExpirationDate": null,
-    "exemptDescription": null,
-    "exemptEntityUseCode": null,
-    "companyCode": null,
-    "VATId": null
-  },
-  "success": true
-}
