+++ handlers/holiday-stop-processor/src/main/scala/com/gu/holidaystopprocessor/Handler.scala
@@ -2,7 +2,7 @@ package com.gu.holidaystopprocessor
 
 import java.time.LocalDate
 import cats.syntax.all._
-import com.amazonaws.services.lambda.runtime.{Context, LambdaRuntime}
+import com.amazonaws.services.lambda.runtime.Context
 import com.gu.creditprocessor.ProcessResult
 import com.gu.effects.GetFromS3
 import com.gu.holiday_stops.{Configuration, ConfigurationLive}
@@ -15,21 +15,9 @@ import io.github.mkotsur.aws.handler.Lambda._
 import zio._
 import zio.console.Console
 
-import java.io.{IOException, OutputStream, PrintStream}
-import java.lang.{System => JavaSystem}
+import java.io.IOException
 
 object Handler extends Lambda[Option[ProductTypeAndStopDate], List[ZuoraHolidayCreditAddResult]] with zio.App {
-
-  val printStream = new PrintStream(new OutputStream() {
-    override def write(b: Int): Unit =
-      LambdaRuntime.getLogger.log(Array(b.toByte))
-
-    override def write(b: Array[Byte]): Unit =
-      LambdaRuntime.getLogger.log(b)
-
-    override def write(b: Array[Byte], off: Int, len: Int): Unit =
-      LambdaRuntime.getLogger.log(b.slice(off, off + len))
-  })
 
   /** @param productTypeAndStopDateOverride
     *   Optionally, the handler will take a product type and stopped publication date of holiday-stop requests to
@@ -40,8 +28,6 @@ object Handler extends Lambda[Option[ProductTypeAndStopDate], List[ZuoraHolidayC
       productTypeAndStopDateOverride: Option[ProductTypeAndStopDate],
       context: Context,
   ): Either[Throwable, List[ZuoraHolidayCreditAddResult]] = {
-    JavaSystem.setOut(printStream)
-    JavaSystem.setErr(printStream)
 
     val runtime = zio.Runtime.default
