+++ handlers/product-move-api/src/main/scala/com/gu/productmove/framework/ZIOApiGatewayRequestHandler.scala
@@ -37,6 +37,9 @@ abstract class ZIOApiGatewayRequestHandler(val server: List[ServerEndpoint[Any, 
 
     override def write(b: Array[Byte], off: Int, len: Int): Unit =
       LambdaRuntime.getLogger.log(b.slice(off, off + len))
+
+    override def flush(): Unit =
+      JavaSystem.out.flush()
   })
 
   // this is the main lambda entry point.  It is referenced in the cloudformation.
