+++ lib/zuora-core/src/main/scala/com/gu/zuora/ZuoraProductTypes.scala
@@ -9,5 +9,4 @@ object ZuoraProductTypes {
   object NewspaperDigitalVoucher extends ZuoraProductType("Newspaper - Digital Voucher")
   object NewspaperNationalDelivery extends ZuoraProductType("Newspaper - National Delivery")
   object GuardianWeekly extends ZuoraProductType("Guardian Weekly")
-  object TierThree extends ZuoraProductType("Tier Three") // Tier Three is a bundle product containing Guardian Weekly
 }
